[AMD] Fix e4m3fn→fnuz NaN corruption in FBGEMMFp8, QuarkW8A8Fp8, and Fp8LinearMethod AITER per-token default
- fpgemm_fp8.py, quark_w8a8_fp8.py: allocate weight buffer as
  float8_e4m3fn (not fp8_dtype=fnuz on AMD) so checkpoint bytes are
  loaded via bitwise copy.  normalize_e4m3fn_to_e4m3fnuz in
  process_weights_after_loading then safely reinterprets the bits as
  fnuz and doubles the scale.  Allocating as fnuz causes copy_ to
  silently NaN-ify values in [240, 448] from NVIDIA FP8 checkpoints.

- fp8.py (Fp8LinearMethod): default use_aiter_fp8_per_token to True
  on AMD when SGLANG_USE_AITER is set and the env var
  SGLANG_USE_AITER_FP8_PER_TOKEN is not explicitly set to 0.
  Per-token×per-channel GEMM (gemm_a8w8_bpreshuffle) is faster on
  MI300X than per-tensor quant for fp8 linear layers.

+++ python/sglang/srt/layers/quantization/fpgemm_fp8.py
@@ -128,11 +128,17 @@ class FBGEMMFp8LinearMethod(LinearMethodBase):
         layer.orig_dtype = params_dtype
 
         # WEIGHT
+        # Always allocate as float8_e4m3fn so checkpoint bytes are loaded via
+        # bitwise copy (no value conversion).  On AMD we normalize to
+        # float8_e4m3fnuz in process_weights_after_loading, which correctly
+        # handles the NaN-bit pattern and doubles the scale.  Allocating as
+        # float8_e4m3fnuz directly would cause values 240-448 to become NaN
+        # and would break the assert in normalize_e4m3fn_to_e4m3fnuz.
         weight = ModelWeightParameter(
             data=torch.empty(
                 output_size_per_partition,
                 input_size_per_partition,
-                dtype=fp8_dtype,
+                dtype=torch.float8_e4m3fn,
             ),
             input_dim=1,
             output_dim=0,
